+++ src/js/view/room/modal/library.mjs
@@ -52,6 +52,9 @@ import {
 } from '../../../state/index.mjs'
 
 import {
+  LAYER_TILE,
+  LAYER_OVERLAY,
+  LAYER_TOKEN,
   createPieceFromAsset,
   populatePieceDefaults,
   splitAssetFilename,
@@ -330,7 +333,7 @@ function modalUpload () {
     }
 
     // set bg color
-    if (type === 'token' || data.format === 'jpg') {
+    if (type === LAYER_TOKEN || data.format === 'jpg') {
       data.bg = _('#upload-color').value
     } else {
       data.bg = 'transparent'
@@ -342,13 +345,13 @@ function modalUpload () {
           .then(() => {
             refreshTabs()
             switch (data.type) {
-              case 'tile':
+              case LAYER_TILE:
                 _('#tab-1').checked = true
                 break
-              case 'overlay':
+              case LAYER_OVERLAY:
                 _('#tab-2').checked = true
                 break
-              case 'token':
+              case LAYER_TOKEN:
                 _('#tab-3').checked = true
                 break
               default:
@@ -398,10 +401,10 @@ function uploadFailed (why) {
 function setupTabUpload () {
   // width
   const type = _('#upload-type')
-  for (const l of ['token', 'overlay', 'tile']) {
+  for (const l of [LAYER_TOKEN, LAYER_OVERLAY, LAYER_TILE]) {
     const option = _('option').create(toTitleCase(l))
     option.value = l
-    if (l === 'token') option.selected = true
+    if (l === LAYER_TOKEN) option.selected = true
     type.add(option)
   }
   type.on('change', change => updatePreview())
@@ -469,7 +472,7 @@ function updatePreview (parseImage = false) {
 
   // add piece to DOM
   const piece = _(`.piece.piece-${type}.is-w-${w}.is-h-${h}`).create()
-  if (type === 'overlay' || type === 'tile') {
+  if (type === LAYER_OVERLAY || type === LAYER_TILE) {
     piece.css({ '--fbg-color': 'rgba(0,0,0,.05)' })
   } else {
     piece.css({ '--fbg-color': '#202020' })