+++ CHANGELOG.md
@@ -15,10 +15,11 @@ Download current and previous versions from [https://github.com/ludus-leonis/Fre
 
 ### Other changes
 
-* fixed dragging some pieces should not change z (e.g. dicemat)
 * added Docker build
 * added multicolor / colormask assets
 * added random material offsets
+* fixed dragging some pieces should not change z (e.g. dicemat)
+* fixed disable-rotate for other-pieces
 
 
 ## v0.14.0 - Verbose Vicuna